+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Configure Spotless' Java extension with the ZAP license.
+- Configure Spotless' Java extension with the ZAP license and Google Java Format (AOSP).
 - Configure the `JavaCompile` tasks to use UTF-8, enable all warnings, and handle warnings as errors.
 
 ### Changed